reset src folder

+++ src/lib/sync/server_observer.ts
@@ -1,6 +1,6 @@
 import ServerNotificationManager from "@/server_notification_manager";
 import { UserOrg } from "@/state/models";
-import { AtuinStore } from "@/state/store";
+import { AtuinStore, useStore } from "@/state/store";
 import { SharedStateManager } from "../shared_state/manager";
 import { AtuinSharedStateAdapter } from "../shared_state/adapter";
 import WorkspaceFolder, { Folder } from "@/state/runbooks/workspace_folders";
@@ -14,10 +14,12 @@ import Runbook from "@/state/runbooks/runbook";
 import { ConnectionState } from "@/state/store/user_state";
 import { autobind } from "../decorators";
 import AppBus from "../app/app_bus";
+import { TabUri } from "@/state/store/ui_state";
 
 const USER_SYNC_INTERVAL = 1000 * 60 * 1;
 
-const queue = new AsyncQueue(5);
+const concurrency = useStore.getState().syncConcurrency;
+const queue = new AsyncQueue(concurrency);
 
 export default class ServerObserver {
   private readonly store: AtuinStore;
@@ -162,8 +164,8 @@ class PersonalOrgObserver {
   }
 
   private async init() {
-    const workspaces = await Workspace.all({ orgId: null });
-    for (const workspace of workspaces) {
+    const personalOnlineWorkspaces = await Workspace.all({ orgId: null, online: 1 });
+    for (const workspace of personalOnlineWorkspaces) {
       this.workspaceObservers.set(
         workspace.get("id")!,
         new OrgWorkspaceObserver(workspace.get("id")!, this.store, this.notifications),
@@ -238,7 +240,7 @@ class OrgObserver {
   }
 
   private async init() {
-    const workspaces = await Workspace.all({ orgId: this.orgId });
+    const workspaces = await Workspace.all({ orgId: this.orgId, online: 1 });
     for (const workspace of workspaces) {
       this.workspaceObservers.set(
         workspace.get("id")!,
@@ -369,13 +371,18 @@ class OrgWorkspaceObserver {
 
   private async deleteRunbook(runbookId: string) {
     const runbook = await Runbook.load(runbookId);
+    if (runbook && runbook.workspaceId !== this.workspaceId) {
+      return;
+    }
+
     if (runbook) {
       await runbook.delete();
     }
 
-    if (this.store.getState().currentRunbookId === runbookId) {
-      this.store.setState({ currentRunbookId: null });
-    }
+    this.store.getState().closeTabs((tab) => {
+      const uri = new TabUri(tab.url);
+      return uri.isRunbook() && uri.getRunbookId() === runbookId;
+    });
   }
 }
 
@@ -460,6 +467,11 @@ class OrgRunbookObserver {
   }
 
   private async syncRunbook(priority: number = 0) {
+    if (!useStore.getState().backgroundSync) {
+      this.logger.debug("Background sync is disabled, skipping");
+      return;
+    }
+
     if (this.syncing) {
       if (!this.isShutdown) {
         this.syncAfterSync = true;
@@ -488,7 +500,13 @@ class OrgRunbookObserver {
     try {
       const user = this.store.getState().user;
       const rbs = new RunbookSynchronizer(this.runbookId, this.workspaceId, user);
-      const result = await rbs.sync(this.runbookId !== this.store.getState().currentRunbookId);
+
+      const tabsMatch = this.store.getState().tabs.some((tab) => {
+        const uri = new TabUri(tab.url);
+        return uri.isRunbook() && uri.getRunbookId() === this.runbookId;
+      });
+
+      const result = await rbs.sync(!tabsMatch);
       if (result.action === "created") {
         // Hack to reset the editor, including the Phoenix Provider
         AppBus.get().emitResetEditor(this.runbookId);
